Add SetSourceResolution method that has the same effect as the --dpi flag
Tesseract uses the source image dpi info if available and otherwise falls
back to 70dpi and prints a warning to stderr. Tesseract works find with
70dpi and only uses that info for recognising font size. However, when
manually preparing and cutting many text blocks using opencv, which does
not encode the dpi in images for recognition the warning messages sure are
annoying and eat performance.